#696: generalise form parsing so it can be used for serverless funcs
