+++ src/Listener/PodeHttpRequest.cs
@@ -378,106 +378,7 @@ namespace Pode
 
         public void ParseFormData()
         {
-            Form = new PodeForm();
-
-            if (RawBody.Length == 0)
-            {
-                return;
-            }
-
-            var lines = PodeHelpers.ConvertToByteLines(RawBody);
-
-            var parts = ContentType.Split(';');
-            var boundaryStart = $"--{parts[1].Split('=')[1].Trim()}";
-            var boundaryEnd = $"{boundaryStart}--";
-
-            var boundaryLineIndexes = new List<int>();
-            for (var i = 0; i < lines.Count; i++)
-            {
-                if (IsLineBoundary(lines[i], boundaryStart) || IsLineBoundary(lines[i], boundaryEnd))
-                {
-                    boundaryLineIndexes.Add(i);
-                }
-            }
-
-            var boundaryLineIndex = 0;
-            var disposition = string.Empty;
-            var fields = new Dictionary<string, string>();
-
-            for (var i = 0; i < (boundaryLineIndexes.Count - 1); i++)
-            {
-                fields.Clear();
-
-                boundaryLineIndex = boundaryLineIndexes[i];
-                disposition = ContentEncoding.GetString(lines[boundaryLineIndex + 1]).Trim(PodeHelpers.NEW_LINE_ARRAY);
-
-                foreach (var line in disposition.Split(';'))
-                {
-                    var atoms = line.Split('=');
-                    if (atoms.Length == 2)
-                    {
-                        fields.Add(atoms[0].Trim(), atoms[1].Trim(' ', '"'));
-                    }
-                }
-
-                if (!fields.ContainsKey("filename"))
-                {
-                    Form.Data.Add(new PodeFormData(fields["name"], ContentEncoding.GetString(lines[boundaryLineIndex + 3]).Trim(PodeHelpers.NEW_LINE_ARRAY)));
-                }
-
-                if (fields.ContainsKey("filename"))
-                {
-                    Form.Data.Add(new PodeFormData(fields["name"], fields["filename"]));
-
-                    if (!string.IsNullOrWhiteSpace(fields["filename"]))
-                    {
-                        var contentType = ContentEncoding.GetString(lines[boundaryLineIndex + 2]).Trim(PodeHelpers.NEW_LINE_ARRAY);
-                        var fileBytesLength = 0;
-                        var stream = new MemoryStream();
-
-                        for (var j = (boundaryLineIndex + 4); j <= (boundaryLineIndexes[i + 1] - 1); j++)
-                        {
-                            fileBytesLength = lines[j].Length;
-                            if (j == (boundaryLineIndexes[i + 1] - 1))
-                            {
-                                if (lines[j][fileBytesLength - 1] == PodeHelpers.NEW_LINE_BYTE)
-                                {
-                                    fileBytesLength--;
-                                }
-
-                                if (lines[j][fileBytesLength - 1] == PodeHelpers.CARRIAGE_RETURN_BYTE)
-                                {
-                                    fileBytesLength--;
-                                }
-                            }
-
-                            stream.Write(lines[j], 0, fileBytesLength);
-                        }
-
-                        Form.Files.Add(new PodeFormFile(fields["filename"], stream, fields["name"], contentType.Split(':')[1].Trim()));
-                    }
-                }
-            }
-        }
-
-        private bool IsLineBoundary(byte[] bytes, string boundary)
-        {
-            if (bytes.Length == 0)
-            {
-                return false;
-            }
-
-            if (bytes[0] != PodeHelpers.DASH_BYTE && bytes[bytes.Length - 1] != PodeHelpers.DASH_BYTE)
-            {
-                return false;
-            }
-
-            if ((bytes.Length - boundary.Length) > 3)
-            {
-                return false;
-            }
-
-            return (ContentEncoding.GetString(bytes).StartsWith(boundary));
+            Form = PodeForm.Parse(RawBody, ContentType, ContentEncoding);
         }
 
         public override void Dispose()
